fix(design): relocate Magic UI to magicui/, restore concentric radii
Phase 1 subagent placed Magic UI components in src/components/ui/
(shadcn's default alias). Moving to src/components/magicui/ per
DESIGN.md §7.1 so our ui/ stays reserved for our own primitives
(button, tech-chip, tech-node, brand-mark, alert, kbd) and later
phases' hardcoded @/components/magicui/* imports resolve.

Also surgical globals.css cleanup — shadcn init's @theme inline
block was silently redefining --radius-sm..2xl via calc(var(--radius)*n),
clobbering our explicit 4/8/12/16/24/32px concentric scale. Same pattern
on --border (oklch) overriding our rgba surface token in :root and .dark.
Removed those 7 conflicting lines; kept the additive --radius-3xl/4xl and
all shadcn primitive tokens (--primary, --background, etc).

StylesPreview imports updated to match new paths.
