+++ src/components/magicui/blur-fade.tsx
