+++ src/components/magicui/marquee.tsx
