+++ src/components/magicui/number-ticker.tsx
